chore: rename BaseMatrix as WrapperMatrix2D
